move `RegExp.escape` to stable ES
