+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 ##### Unreleased
+- [`RegExp.escape` proposal](https://github.com/tc39/proposal-regex-escaping):
+  - Built-ins:
+    - `RegExp.escape`
+  - Moved to stable ES, February 2025 TC39 meeting
+  - Added `es.` namespace module, `/es/` and `/stable/` namespaces entries
 - Compat data improvements:
   - [`DisposableStack`, `AsyncDisposableStack`, `SuppressedError` and `Iterator.prototype[@@dispose]`](https://github.com/tc39/proposal-explicit-resource-management) marked as [shipped from V8 ~ Chromium 134](https://issues.chromium.org/issues/42203506#comment24)
   - [`Error.isError`](https://github.com/tc39/proposal-is-error) added and marked as supported from V8 ~ Chromium 134
