+++ packages/core-js/es/regexp/escape.js
@@ -1,0 +1,5 @@
+'use strict';
+require('../../modules/es.regexp.escape');
+var path = require('../../internals/path');
+
+module.exports = path.RegExp.escape;
