+++ packages/core-js/stable/regexp/escape.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/regexp/escape');
+
+module.exports = parent;
