+++ packages/core-js/stage/3.js
@@ -9,7 +9,6 @@ require('../proposals/float16');
 require('../proposals/is-error');
 require('../proposals/json-parse-with-source');
 require('../proposals/math-sum');
-require('../proposals/regexp-escaping');
 // TODO: Obsolete versions, remove from `core-js@4`
 require('../proposals/array-grouping-stage-3');
 require('../proposals/array-grouping-stage-3-2');
