+++ tests/unit-global/es.regexp.escape.js
